Check all shell scripts with shellcheck

+++ assets/chezmoi.io/make-gh-pages.sh
@@ -7,21 +7,20 @@ rm -rf public/
 hugo
 
 # clone and checkout the gh-pages branch in a temporary directory
-oldwd=$(pwd)
 tmpdir=$(mktemp -d)
 cleanup() {
-    rm -rf ${tmpdir}
+    rm -rf "${tmpdir}"
 }
 trap cleanup EXIT
 git branch -f gh-pages origin/gh-pages
-git clone --branch=gh-pages --local ../.. ${tmpdir}
+git clone --branch=gh-pages --local ../.. "${tmpdir}"
 
 # copy the new website to the temporary directory
-rm -rf ${tmpdir}/*
-cp -r public/* ${tmpdir}
+rm -rf "${tmpdir:?}"/*
+cp -r public/* "${tmpdir}"
 
 # prepare the clone
-cd ${tmpdir}
+cd "${tmpdir}"
 git checkout CNAME
 git remote set-url origin https://github.com/twpayne/chezmoi.git
 git fetch origin
